requirements_production.txt
# Core FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Task Queue
celery==5.3.6
redis==5.0.1

# Object Storage (S3/MinIO)
boto3==1.34.34

# Video Processing
opencv-python-headless==4.9.0.80
pillow==10.2.0

# AI/ML
torch==2.2.0
transformers==4.37.2
accelerate==0.26.1
bitsandbytes==0.42.0

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.26.0

# Monitoring & Logging
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.40.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
httpx==0.26.0

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0
